auth-backend: split OAuthProvider into lib/oauth and lib/flow
