+++ plugins/auth-backend/src/providers/microsoft/provider.ts
@@ -35,7 +35,7 @@ import {
   PassportDoneCallback,
 } from '../types';
 
-import { OAuthProvider } from '../../lib/OAuthProvider';
+import { OAuthProvider } from '../../lib/oauth';
 import { Logger } from 'winston';
 import { TokenIssuer } from '../../identity';
 import { Config } from '@backstage/config';
